==> notebooks/1.1_separate_CBS-QB3_G4_calculations.ipynb <==
import sys
sys.path.insert(0,"/home/gridsan/hwpang/Software/RMG-Py/")
sys.path.insert(0,"..")

from rmgpy.data.thermo import ThermoLibrary, ThermoDatabase

thermo_library = ThermoLibrary()
thermo_library.load(
    "../data/johnson.py",
    ThermoDatabase().local_context,
    ThermoDatabase().global_context,
)

thermo_library_cbsqb3 = ThermoLibrary()

for label, entry in list(thermo_library.entries.items()):
    if "CBS-QB3" in entry.long_desc:
        thermo_library_cbsqb3.entries[label] = entry
        del thermo_library.entries[label]

thermo_library_cbsqb3.save("../data/johnson_cbs-qb3.py")

thermo_library.save("../data/johnson_g4.py")
