==> notebooks/2.1_get_radical_data.ipynb <==
# Get collected radical data

import sys
sys.path.insert(0,"/home/gridsan/hwpang/Software/RMG-Py/")
sys.path.insert(0,"..")

import random
import os
import time
import math
from copy import deepcopy
import pandas as pd
from joblib import Parallel, delayed
from tqdm import tqdm

from rmgpy.data.thermo import ThermoDatabase, ThermoLibrary, ThermoData, remove_thermo_data, add_thermo_data, NASA
from rmgpy.molecule import Molecule
from rmgpy.species import Species
from rmgpy import settings
from rmgpy import constants

from tree.utils import load_thermo_lib_by_path
from tree.parameters import Ts

# Extract radical data into DataFrame

lib_paths = [
    "../data/dong_pio_liang.py",
    "../data/pang.py",
    "../data/johnson_cbs-qb3.py",
    "../data/johnson_g4.py",
]

# load thermo database

thermo_database = ThermoDatabase()
thermo_database.load_groups(os.path.join(settings["database.directory"], "thermo", "groups"))
for path in lib_paths:
    load_thermo_lib_by_path(path, thermo_database)
    thermo_database.library_order.remove(path)
thermo_database.library_order = lib_paths + thermo_database.library_order

temp_dict = {
    "smiles": [],
    "H298 (kcal/mol)": [],
    "S298 (cal/mol/K)": [],
    "Sint298 (cal/mol/K)": [],
    "source": [],
    "level_of_theory": []
}

for T in Ts:
    temp_dict[f"Cp{T} (cal/mol/K)"] = []

for lib_path in lib_paths:
    lib = thermo_database.libraries[lib_path]
    for entry in lib.entries.values():
        if sum(atom.radical_electrons for atom in entry.item.atoms)==1 and any(not atom.is_hydrogen() and not atom.is_halogen() for atom in entry.item.atoms):
            radical = deepcopy(entry.item)
            radical_data = deepcopy(entry.data)

            temp_dict["smiles"].append(radical.to_smiles())
            temp_dict["source"].append(os.path.basename(lib_path))
            if lib_path == "../data/johnson_g4.py":
                    level_of_theory = "G4"
            else:
                level_of_theory = "CBS-QB3"
            temp_dict["level_of_theory"].append(level_of_theory)
            
            radical_data.Cp0 = (radical.calculate_cp0(), "J/(mol*K)")
            radical_data.CpInf = (radical.calculate_cpinf(), "J/(mol*K)")
            if not isinstance(radical_data, ThermoData):
                radical_data = radical_data.to_thermo_data()
                
            temp_dict["H298 (kcal/mol)"].append(radical_data.H298.value_si/1000/4.184)
            temp_dict["S298 (cal/mol/K)"].append(radical_data.S298.value_si/4.184)
            for Cp, T in zip(radical_data.Cpdata.value_si, Ts):
                temp_dict[f"Cp{T} (cal/mol/K)"].append(Cp/4.184)
            
            #remove entropy correction
            radical_data.S298.value_si += constants.R * math.log(radical.get_symmetry_number())
            temp_dict["Sint298 (cal/mol/K)"].append(radical_data.S298.value_si/4.184)
            
radical_df = pd.DataFrame(temp_dict)
radical_df

radical_df = radical_df.sort_values("H298 (kcal/mol)", ascending=True)
radical_df

radical_df = radical_df.drop_duplicates(subset=["smiles"], keep="first")
radical_df = radical_df.reset_index(drop=True)
radical_df

radical_df.to_csv("../data/radical_data.csv", index=False)
