==> notebooks/2.2_get_resonance_radicals_and_descriptors.ipynb <==
# Get resonance radicals

import sys
sys.path.insert(0,"/home/gridsan/hwpang/Software/RMG-Py/")
sys.path.insert(0, "..")

import random
import os
import time
import math
from copy import deepcopy
import pandas as pd
from joblib import Parallel, delayed
from tqdm import tqdm

from rdkit.Chem.Descriptors import ExactMolWt, NumRadicalElectrons
from rdkit.Chem.rdMolDescriptors import CalcNumRings, CalcNumHBD, CalcNumHeavyAtoms, CalcNumRotatableBonds
from rdkit.Chem import GetPeriodicTable

from rdmc.mol import RDKitMol

from rmgpy.data.thermo import ThermoDatabase, ThermoLibrary, ThermoData, remove_thermo_data, add_thermo_data, NASA
from rmgpy.molecule import Molecule
from rmgpy.species import Species
from rmgpy import settings
from rmgpy import constants

from tree.utils import load_thermo_lib_by_path, generate_thermo
from tree.parameters import Ts

# Generate resonance structures

radical_df = pd.read_csv("../data/radical_data.csv")
radical_df

def generate_radical_resonance_smis(smi):
    rad = Molecule().from_smiles(smi)
    rads = [r for r in rad.generate_resonance_structures() if sum(atom.radical_electrons for atom in r.atoms)==1]
    for r in rads:
        r.atoms.sort()
    return [r.to_smiles() for r in rads]
    
generate_radical_resonance_smis("C[CH]C(=O)OCO")

# takes a min
resonance_smiss = [generate_radical_resonance_smis(smi) for smi in tqdm(radical_df["smiles"])]

temp_dict = {
    "radical_smiles": [],
    "resonance_radical_smiles": [],
}

for smi, resonance_smis in zip(radical_df["smiles"], resonance_smiss):
    for resonance_smi in resonance_smis:
        temp_dict["radical_smiles"].append(smi)
        temp_dict["resonance_radical_smiles"].append(resonance_smi)

resonance_radical_df = pd.DataFrame(temp_dict)
resonance_radical_df

resonance_radical_df = resonance_radical_df.drop_duplicates(subset=["resonance_radical_smiles"], keep="first")
resonance_radical_df = resonance_radical_df.reset_index(drop=True)
resonance_radical_df

resonance_radical_df.to_csv("../data/resonance_radicals.csv", index=False)

# Get descriptors

resonance_radical_df = pd.read_csv("../data/resonance_radicals.csv")

def get_descriptors(smi):
    mol = RDKitMol.FromSmiles(smi)
    radical_atom = [atom for atom in mol.GetAtoms() if atom.GetNumRadicalElectrons()!=0][0]
    return set(mol.GetElementSymbols()), ExactMolWt(mol._mol), CalcNumRings(mol._mol), CalcNumHeavyAtoms(mol._mol), CalcNumRotatableBonds(mol._mol), radical_atom.GetSymbol()

get_descriptors("[CH3]")

descriptorss = Parallel(backend="multiprocessing", n_jobs=4)(delayed(get_descriptors)(smi) for smi in tqdm(resonance_radical_df["resonance_radical_smiles"]))




atom_symbolss = [descriptors[0] for descriptors in descriptorss]

all_atom_symbols = set()
for atom_symbols in atom_symbolss:
    all_atom_symbols.update(atom_symbols)
all_atom_symbols

PeriodicTable = GetPeriodicTable()
sorted_all_atom_symbols = list(all_atom_symbols)
sorted_all_atom_symbols.sort(key=lambda x: PeriodicTable.GetAtomicNumber(x))
sorted_all_atom_symbols

for atom_symbol in sorted_all_atom_symbols:
    resonance_radical_df[f"has_{atom_symbol}"] = [atom_symbol in atom_symbols for atom_symbols in atom_symbolss]
resonance_radical_df

resonance_radical_df["Mw (g/mol)"] = [descriptors[1] for descriptors in descriptorss]

resonance_radical_df["num_rings"] = [descriptors[2] for descriptors in descriptorss]

resonance_radical_df["num_heavy_atoms"] = [descriptors[3] for descriptors in descriptorss]

resonance_radical_df["num_rotatable_bonds"] = [descriptors[4] for descriptors in descriptorss]

resonance_radical_df["radical_atom_type"] = [descriptors[5] for descriptors in descriptorss]

resonance_radical_df

resonance_radical_df.to_csv("../data/resonance_radicals.csv", index=False)

# Drop duplicates in radical data

radical_data_df = pd.read_csv("../data/radical_data.csv")
radical_data_df

resonance_radical_df = pd.read_csv("../data/resonance_radicals.csv")
resonance_radical_df

radical_data_df = radical_data_df[radical_data_df["smiles"].isin(resonance_radical_df["radical_smiles"])]
radical_data_df

radical_data_df = radical_data_df.reset_index(drop=True)
radical_data_df

radical_data_df.to_csv("../data/radical_data_no_dup.csv", index=False)
