==> notebooks/2.4_get_GAV_estimation_for_radical_data.ipynb <==
# Get GAV estimation for collected radical data

import sys
sys.path.insert(0,"/home/gridsan/hwpang/Software/RMG-Py/")
sys.path.insert(0,"..")

import random
import os
import time
import math
from copy import deepcopy
import pandas as pd
from joblib import Parallel, delayed
from tqdm import tqdm

from rmgpy.data.thermo import ThermoDatabase, ThermoLibrary, ThermoData, remove_thermo_data, add_thermo_data, NASA
from rmgpy.molecule import Molecule
from rmgpy.species import Species
from rmgpy import settings
from rmgpy import constants

from tree.utils import load_thermo_lib_by_path, generate_thermo
from tree.parameters import Ts

# Load radical data

path = "../data/radical_data_no_dup.csv"
radical_data_df = pd.read_csv(path)
radical_data_df

# Estimate GAV for radicals

# load thermo database

thermo_database = ThermoDatabase()
thermo_database.load_groups(os.path.join(settings["database.directory"], "thermo", "groups"))

generate_thermo(thermo_database, "[O]C(=O)OC(O)(O)O", resonance=True) # we want the best thermo estimation for the radical

radical_gav_df = radical_data_df[["smiles"]]
thermos = radical_gav_df["smiles"].apply(lambda smi: generate_thermo(thermo_database, smi, resonance=True))

columns = ["H298 (kcal/mol)", "S298 (cal/mol/K)", "Cp (cal/mol/K)", "Sint298 (cal/mol/K)", "comment"]
radical_gav_df[columns] = thermos.to_list()

columns = []
for i, T in enumerate(Ts):
    columns.append(f"Cp{T} (cal/mol/K)")
radical_gav_df[columns] = radical_gav_df["Cp (cal/mol/K)"].to_list()

radical_gav_df = radical_gav_df.drop("Cp (cal/mol/K)", axis=1)

columns = radical_gav_df.columns.tolist()
columns.remove("comment")
columns.append("comment")

radical_gav_df = radical_gav_df[columns]

radical_gav_df

radical_gav_df.to_csv("../data/radical_gav.csv", index=False)
