==> notebooks/3.1_get_GAV_estimation_for_resoance_radical.ipynb <==
# Get GAV estimation for resonance radical

import sys
sys.path.insert(0,"/home/gridsan/hwpang/Software/RMG-Py/")
sys.path.insert(0,"..")

import random
import os
import time
import math
from copy import deepcopy
import pandas as pd
from joblib import Parallel, delayed
from tqdm import tqdm

from rmgpy.data.thermo import ThermoDatabase, ThermoLibrary, ThermoData, remove_thermo_data, add_thermo_data, NASA, ThermoGroups
from rmgpy.molecule import Molecule
from rmgpy.species import Species
from rmgpy import settings
from rmgpy import constants


from tree.utils import load_thermo_lib_by_path, generate_thermo
from tree.parameters import Ts

# Load resonance radicals

resonance_radical_df = pd.read_csv("../data/resonance_radicals.csv")
resonance_radical_df

# Estimate GAV for resonance radicals

# load thermo database

thermo_database = ThermoDatabase()
thermo_database.load_groups(os.path.join(settings["database.directory"], "thermo", "groups"))

generate_thermo(thermo_database, "[O]C(=O)OC(O)(O)O", resonance=False) # we want to know the GAV estimation for the exact resonance radical

resonance_radical_gav_df = pd.DataFrame()
resonance_radical_gav_df["smiles"] = resonance_radical_df["resonance_radical_smiles"]

thermos = resonance_radical_gav_df["smiles"].apply(lambda smi: generate_thermo(thermo_database, smi, resonance=False))

temp_columns = ["H298 (kcal/mol)", "S298 (cal/mol/K)", "Cp (cal/mol/K)", "Sint298 (cal/mol/K)", "comment"]
resonance_radical_gav_df[temp_columns] = thermos.to_list()

temp_columns = []
for i, T in enumerate(Ts):
    temp_columns.append(f"Cp{T} (cal/mol/K)")
resonance_radical_gav_df[temp_columns] = resonance_radical_gav_df["Cp (cal/mol/K)"].to_list()

resonance_radical_gav_df = resonance_radical_gav_df.drop("Cp (cal/mol/K)", axis=1)

temp_columns = resonance_radical_gav_df.columns.tolist()
temp_columns.remove("comment")
temp_columns.append("comment")

resonance_radical_gav_df = resonance_radical_gav_df[temp_columns]
resonance_radical_gav_df

resonance_radical_gav_df.to_csv("../data/resonance_radical_gav.csv", index=False)
