==> notebooks/4.2_get_corresponding_closed_shells_and_data.ipynb <==
# Get corresponding closed shells and data

import sys
sys.path.insert(0,"/home/gridsan/hwpang/Software/RMG-Py/")
sys.path.insert(0,"..")

import random
import os
import time
import math
from copy import deepcopy
import pandas as pd
from joblib import Parallel, delayed
from tqdm import tqdm

from rmgpy.data.thermo import ThermoDatabase, ThermoLibrary, ThermoData, remove_thermo_data, add_thermo_data, NASA
from rmgpy.molecule import Molecule
from rmgpy.species import Species
from rmgpy import settings
from rmgpy import constants

from tree.utils import load_thermo_lib_by_path, generate_thermo
from tree.parameters import Ts

# Load resonance radicals

resonance_radical_df = pd.read_csv("../data/resonance_radicals.csv")

# Generate closed shell structures

closed_shell_df = resonance_radical_df[["resonance_radical_smiles"]]

def generate_closed_shell_smi(smi):
    mol = Molecule().from_smiles(smi)
    mol.saturate_radicals()
    mol.atoms.sort()
    return mol.to_smiles()

generate_closed_shell_smi("CC1OC1OC[O]")

closed_shell_df["closed_shell_smiles"] = closed_shell_df["resonance_radical_smiles"].apply(generate_closed_shell_smi)
closed_shell_df

closed_shell_df.to_csv("../data/closed_shells.csv", index=False)

# Generate closed shell thermo

lib_paths = [
    "../data/dong_pio_liang.py",
    "../data/pang.py",
    "../data/johnson_cbs-qb3.py",
    "../data/johnson_g4.py",
]

# load thermo database

thermo_database = ThermoDatabase()
thermo_database.load_groups(os.path.join(settings["database.directory"], "thermo", "groups"))
for path in lib_paths:
    load_thermo_lib_by_path(path, thermo_database)
    thermo_database.library_order.remove(path)
thermo_database.library_order = lib_paths + thermo_database.library_order

# takes ~1 min 30 sec
closed_shell_thermos = [generate_thermo(thermo_database, smi, resonance=False) for smi in tqdm(closed_shell_df["closed_shell_smiles"])]


closed_shell_df["closed_shell_H298 (kcal/mol)"] = [thermo[0] for thermo in closed_shell_thermos]
closed_shell_df["closed_shell_S298 (cal/mol/K)"] = [thermo[1] for thermo in closed_shell_thermos]
for i, T in enumerate(Ts):
    closed_shell_df[f"closed_shell_Cp{T} (cal/mol/K)"] = [thermo[2][i] for thermo in closed_shell_thermos]
closed_shell_df["closed_shell_Sint298 (cal/mol/K)"] = [thermo[3] for thermo in closed_shell_thermos]
closed_shell_df["closed_shell_thermo_comment"] = [thermo[4] for thermo in closed_shell_thermos]
closed_shell_df

def thermo_source(comment):
    if "Thermo group additivity estimation" in comment:
        return "GAV"
    else:
        return comment
closed_shell_df["closed_shell_thermo_source"] = closed_shell_df["closed_shell_thermo_comment"].apply(thermo_source)
closed_shell_df

closed_shell_df.to_csv("../data/closed_shell_data.csv", index=False)

# Get stats

closed_shell_df["closed_shell_thermo_source"].value_counts()

873+81+5

sum(closed_shell_df["closed_shell_thermo_source"].value_counts())
