==> notebooks/5.1_get_hbis.ipynb <==
# Get HBI data

import sys
sys.path.insert(0,"/home/gridsan/hwpang/Software/RMG-Py/")
sys.path.insert(0,"..")

import random
import os
import time
import math
from copy import deepcopy
import pandas as pd
from joblib import Parallel, delayed
from tqdm import tqdm

from rmgpy.data.thermo import ThermoDatabase, ThermoLibrary, ThermoData, remove_thermo_data, add_thermo_data, NASA
from rmgpy.molecule import Molecule
from rmgpy.species import Species
from rmgpy import settings
from rmgpy import constants

from tree.utils import load_thermo_lib_by_path, generate_thermo
from tree.parameters import Ts

# Load data to join

radical_data_df = pd.read_csv("../data/radical_data_no_dup.csv")
resonance_radical_df = pd.read_csv("../data/resonance_radicals.csv")
closed_shell_data_df = pd.read_csv("../data/closed_shell_data.csv")

resonance_radical_df

radical_data_df

# Join data

temp_radical_data_df = radical_data_df
temp_radical_data_df = temp_radical_data_df.rename(columns={column: "radical_"+column for column in radical_data_df.columns})
temp_radical_data_df = temp_radical_data_df.drop(columns="radical_S298 (cal/mol/K)")
temp_radical_data_df

temp_closed_shell_data_df = closed_shell_data_df.drop(columns="closed_shell_S298 (cal/mol/K)")

temp_closed_shell_data_df["resonance_radical_smiles"].value_counts()

hbi_df = resonance_radical_df[["radical_smiles", "resonance_radical_smiles", "num_rotatable_bonds"]]
hbi_df = hbi_df.rename(columns={"num_rotatable_bonds": "resonance_radical_num_rotatable_bonds"})
hbi_df = hbi_df.merge(temp_radical_data_df, on="radical_smiles", how="left")
hbi_df = hbi_df.merge(temp_closed_shell_data_df, on="resonance_radical_smiles", how="left")
hbi_df

# Derive HBI corrections

properties = ["H298 (kcal/mol)", "Sint298 (cal/mol/K)"]
for T in Ts:
    properties.append(f"Cp{T} (cal/mol/K)")

hbi_df[f"HBI_H298 (kcal/mol)"] = hbi_df[f"radical_H298 (kcal/mol)"] - hbi_df[f"closed_shell_H298 (kcal/mol)"] + 52.1

for prop in properties[1:]:
    hbi_df[f"HBI_{prop}"] = hbi_df[f"radical_{prop}"] - hbi_df[f"closed_shell_{prop}"]
    
hbi_df

hbi_df.to_csv("../data/hbi.csv", index=False)
