==> notebooks/5.2_get_uncertainty_for_hbi_data.ipynb <==
# Get uncertainty for HBI data

import sys
sys.path.insert(0,"/home/gridsan/hwpang/Software/RMG-Py/")
sys.path.insert(0,"..")

import pandas as pd
import numpy as np

from tree.parameters import columns

hbi_df = pd.read_csv("../data/hbi.csv")
hbi_df

resonance_radical_df = pd.read_csv("../data/resonance_radicals.csv")
resonance_radical_df

# Add representative uncertainty

hbi_unc_df = hbi_df

hbi_unc_df.loc[hbi_unc_df["radical_level_of_theory"]=="CBS-QB3", "unc_radical_H298 (kcal/mol)"] = 1.2
hbi_unc_df.loc[hbi_unc_df["radical_level_of_theory"]=="G4", "unc_radical_H298 (kcal/mol)"] = 1.0

temp_columns = ["unc_radical_" + column for column in columns[1:] if "S298" not in column]
print(temp_columns)

hbi_unc_df[temp_columns] = 1.0

hbi_unc_df

temp_sources = [
    "Thermo library: ../data/dong_pio_liang.py",
    "Thermo library: ../data/pang.py",
    "Thermo library: ../data/johnson_cbs-qb3.py",
]
hbi_unc_df.loc[hbi_unc_df["closed_shell_thermo_source"].isin(temp_sources), "unc_closed_shell_H298 (kcal/mol)"] = 1.2
hbi_unc_df.loc[hbi_unc_df["closed_shell_thermo_source"]=="Thermo library: ../data/johnson_g4.py", "unc_closed_shell_H298 (kcal/mol)"] = 1.0
hbi_unc_df.loc[hbi_unc_df["closed_shell_thermo_source"]=="GAV", "unc_closed_shell_H298 (kcal/mol)"] = 4.0

temp_columns = ["unc_closed_shell_" + column for column in columns[1:] if "S298" not in column]

query = hbi_unc_df["closed_shell_thermo_source"]!="GAV"
hbi_unc_df.loc[query, temp_columns] = 1.0
query = hbi_unc_df["closed_shell_thermo_source"]=="GAV"
hbi_unc_df.loc[query, temp_columns] = 2.0

hbi_unc_df

for column in columns:
    if "S298" not in column:
        hbi_unc_df["unc_HBI_" + column] = np.sqrt(hbi_unc_df["unc_radical_" + column]**2 + hbi_unc_df["unc_closed_shell_" + column]**2)
    
hbi_unc_df

hbi_unc_df["unc_HBI_H298 (kcal/mol)"].value_counts()

hbi_unc_df.to_csv("../data/hbi_unc.csv", index=False)
