==> notebooks/6.2_split_data.ipynb <==
# Split data

import json
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans
from scipy.spatial.distance import cdist
from tqdm import tqdm
from rdkit.Chem import rdFingerprintGenerator
import numpy as np
import matplotlib.pyplot as plt
import matplotlib
matplotlib.rcParams['mathtext.fontset'] = 'stix'
matplotlib.rcParams['font.family'] = 'STIXGeneral'

from rdmc.mol import RDKitMol

def make_rdkit_mol(smi):
    return RDKitMol.FromSmiles(smi)._mol

hbi_unc_df = pd.read_csv("../data/hbi_unc.csv")

# Random split

unique_radicals = hbi_unc_df["radical_smiles"].unique()

train, test = train_test_split(unique_radicals, test_size=0.1, random_state=0)

train_df = hbi_unc_df[hbi_unc_df["radical_smiles"].isin(train)]
train_inds = list(train_df.index)

test_df = hbi_unc_df[hbi_unc_df["radical_smiles"].isin(test)]
test_inds = list(test_df.index)

with open("../data/splits/random.json", "w") as f:
    json.dump((train_inds, test_inds), f)

# Random split with validation

unique_radicals = hbi_unc_df["radical_smiles"].unique()
train, test = train_test_split(unique_radicals, test_size=0.1, random_state=0)
train, val = train_test_split(train, test_size=0.1 / 0.9, random_state=0)

train_df = hbi_unc_df[hbi_unc_df["radical_smiles"].isin(train)]
train_inds = list(train_df.index)
len(train_inds)

val_df = hbi_unc_df[hbi_unc_df["radical_smiles"].isin(val)]
val_inds = list(val_df.index)
len(val_inds)

test_df = hbi_unc_df[hbi_unc_df["radical_smiles"].isin(test)]
test_inds = list(test_df.index)
len(test_inds)

with open("../data/splits/random_val.json", "w") as f:
    json.dump((train_inds, val_inds, test_inds), f)

# Cluster split

test_size = 0.1

cluster_df = pd.read_csv("../data/resonance_radical_cluster.csv")

cluster_count_dict = cluster_df["cluster"].value_counts().to_dict()
cluster_count_dict

sorted_clusters = np.array(list(cluster_count_dict.keys())[::-1])
sorted_cluster_counts = np.array(list(cluster_count_dict.values())[::-1])
cluster_cumsums = np.cumsum(sorted_cluster_counts)
cluster_cumsums = cluster_cumsums / cluster_cumsums[-1]
ind = np.argmax(cluster_cumsums >= test_size)
test_clusters = sorted_clusters[:ind+1]
print(test_clusters)
test_smis = cluster_df.loc[cluster_df["cluster"].isin(test_clusters), "resonance_radical_smiles"]
test_smis

test_df = hbi_unc_df[hbi_unc_df["resonance_radical_smiles"].isin(test_smis)]
train_df = hbi_unc_df[~hbi_unc_df.index.isin(test_df.index)]

train_inds = list(train_df.index)
print(len(train_inds)/len(hbi_unc_df.index))
test_inds = list(test_df.index)
print(len(test_inds)/len(hbi_unc_df.index))

with open("../data/splits/cluster.json", "w") as f:
    json.dump((train_inds, test_inds), f)
