==> notebooks/8.2_predict.ipynb <==
# Predict

import sys
sys.path.insert(0,"/home/gridsan/hwpang/Software/RMG-Py/")
sys.path.insert(0,"..")

import json
import pandas as pd

from rmgpy.data.thermo import ThermoDatabase, ThermoGroups

from tree.thermo import ThermoGroups as SIDTThermoGroups
from tree.utils import make_mol
from tree.parameters import Ts

# Load test data

hbi_unc_df = pd.read_csv("../data/hbi_unc.csv")

split = "random"
# split = "cluster"

with open(f"../data/splits/{split}.json", "r") as f:
    train_inds, test_inds = json.load(f)

test_df = hbi_unc_df.loc[test_inds, :]
test_df

mols = test_df["resonance_radical_smiles"].apply(make_mol)
mols = mols.to_list()

# Load thermo database

def make_prediction(thermo_database, tree, mol):
    atoms = {"*": atom for atom in mol.atoms if atom.radical_electrons==1}
    return thermo_database._add_group_thermo_data(None, tree, mol, atoms)[0]

thermo_database = ThermoDatabase()

# Predict with SIDT tree

model_dir = f"../models/split-{split}_run-retrain-empirical"
sidt_tree = SIDTThermoGroups().load(f"{model_dir}/tree.py", thermo_database.local_context, thermo_database.global_context)

thermos = [make_prediction(thermo_database, sidt_tree, mol) for mol in mols]

test_result_df = test_df[["resonance_radical_smiles"]]
test_result_df["HBI_H298 (kcal/mol)"] = [thermo.H298.value_si/4184 for thermo in thermos]
test_result_df["unc_HBI_H298 (kcal/mol)"] = [thermo.H298.uncertainty_si/4184 for thermo in thermos]
test_result_df["HBI_Sint298 (cal/mol/K)"] = [thermo.S298.value_si/4.184 for thermo in thermos]
test_result_df["unc_HBI_Sint298 (cal/mol/K)"] = [thermo.S298.uncertainty_si/4.184 for thermo in thermos]
for i, T in enumerate(Ts):
    test_result_df[f"HBI_Cp{T} (cal/mol/K)"] = [thermo.Cpdata.value_si[i]/4.184 for thermo in thermos]
    test_result_df[f"unc_HBI_Cp{T} (cal/mol/K)"] = [thermo.Cpdata.uncertainty_si[i]/4.184 for thermo in thermos]
test_result_df["comment"] = [thermo.comment for thermo in thermos]

test_result_df.to_csv(f"{model_dir}/test.csv", index=False)
