==> notebooks/8.3_predict_radical.ipynb <==
# Predict radical

import sys
sys.path.insert(0,"/home/gridsan/hwpang/Software/RMG-Py/")
sys.path.insert(0,"..")

import os
import json
import pandas as pd

from rmgpy import settings
from rmgpy.data.thermo import ThermoDatabase, ThermoGroups

from tree.thermo import ThermoGroups as SIDTThermoGroups
from tree.utils import make_mol, generate_thermo
from tree.parameters import Ts

# Load test data

hbi_unc_df = pd.read_csv("../data/hbi_unc.csv")

with open("../data/splits/random.json", "r") as f:
    train_inds, test_inds = json.load(f)

test_df = hbi_unc_df.loc[test_inds, :]
test_df

radical_smis = test_df["radical_smiles"].drop_duplicates(keep="first")

# Load thermo database

# load thermo database

thermo_database = ThermoDatabase()
thermo_database.load_groups(os.path.join(settings["database.directory"], "thermo", "groups"))

# model_dir = "../models/split-random_run-aleatoric-prepruning-bounded/frac-1.0/"
model_dir = "../models/split-random_run-retrain-empirical"
sidt_tree = SIDTThermoGroups().load(f"{model_dir}/tree.py", thermo_database.local_context, thermo_database.global_context)

thermo_database.groups["radical"] = sidt_tree

# Predict with SIDT tree

generate_thermo(thermo_database, "[O]C(=O)OC(O)(O)O", resonance=True) # we want to know the best GAV estimation



thermos = [generate_thermo(thermo_database, smi, resonance=True) for smi in radical_smis]

test_result_df = pd.DataFrame()
test_result_df["radical_smiles"] = radical_smis
test_result_df["H298 (kcal/mol)"] = [thermo[0] for thermo in thermos]
test_result_df["S298 (cal/mol/K)"] = [thermo[1] for thermo in thermos]
for i, T in enumerate(Ts):
    test_result_df[f"Cp{T} (cal/mol/K)"] = [thermo[2][i] for thermo in thermos]
test_result_df["comment"] = [thermo[4] for thermo in thermos]

test_result_df.to_csv(f"{model_dir}/test_radical.csv", index=False)
